* introducing a oyConnector_s to oyFilterNodes_s
The connection is now managed by a oyFilterPlug_s and a opposite
oyFilterSocket_s object. If connected each of them should know the
other. The oyCMMapi4_s backend structure provides prototypes of the
oyConnector_s', to tell a oyFilter_s, which connections are possible.
The oyFilterNode_s on the output side of the graph can check if two
oyConnector_s match. He knows if the associated image can be processed
by the desired connector. If this is successful the oyFilterPlug_s and
oyFilterSocket_s objects can be obtained from the according
oyFilterNode_s objects. The plug will be then connected into the socket
and thus build the graph. There are no more any direct node dependencies
in the graph.

Further the oyOBJECT_TYPE_e was shortened to oyOBJECT_e.

As well some enum alike "defines" where changed from oy to a OY_ prefix,
for consistency.

The Coding Conventions section in the documentation was updated to
clear the Oyranos coding style.