* publish the include directory where Oyranos headers are placed
As discussed recently on the Cairo list, some users prefere a scheme,
where the full path to global includes is to be specified on user side.
The pkg-config and configure scripts are updated to provide that path.
makefile.in is fixed to allow again for make rpm (to some degree).