* [alpha]: add a parant_type_ member to oyStructList_s
This allowes to use the list member of various object lists independent from
the main object. The intention is to use the oyStructList_s generically
in oyOption_s as the new oyValue_u::oy_struct member. Passing of profiles
inside a options set would become trivial and to a good extent transparent.