* [alpha]: simplify oyCMMInfoGetText_f to oyCMMGetText_f
This omits the previously passed static "data" argument and
unifies the string information providing. The select argument
must be agreed. If not within a static agreement then in a
member variable or the like. Three strings is sometimes too less
in oyName_s and does not yet any selecting. oyOption_s is
overkill for backends.