* [oyX1]: add XRandR properties to oyMonitor_s
oyMonitor_newFrom_() is reorganised to do the X detection and select
the a informative API. EDID will be trasnported through a oyBlob_s
element. Currently Xinerama and XRandR code diverge, which is a bug.
The geometry information is separated from the screen information.
The later will be available in case of no more meaningful screen
description like a XRandR output name.