* [alpha]: add oyDevicesList(), oyConfig_EraseFromDB() ...
oyConfig_FromDeviceName() and oyOption_StructMoveIn().
oySetMonitorProfile() is rewritten to use the new API's including
oyConfig_EraseFromDB() but it is not yet tested. oyDevicesList() and
oyConfig_FromDeviceName() is a split out of code from oySetMonitorProfile().
Code organisation will still take some time. But the results are already
visible in a test application.