* [alpha]: change oyOptions_SetFromText() arg
the options argument is passed as pointer to pointer, in order to be able to
create new options with one call. This eleminates the need to check in
advance for existence of the according argument. The design follows
the xxx_MoveIn() and xxx_Release() convention to take over full
responsibility of a argument.