* [alpha]: accept non order edges
The nodes where previously expected to hold edges in a continuous list.
Now edges can be placed depending on positions, which makes adressing easier
as the search for a named connection in a list of active connections (edges)
can be omitted by just using a position parameter. The short come of this
approach is the cost of allocating the number of allowed edges before using
one. This causes overhead. A limitation of permitted edges can keep that
allocation manageable.