* [alpha]: in the fly initialise processing data
oyFilterPlug_ResolveImage() does the check and if needed the graph call,
ticket and socket maintainance. Its good to have in any plug-in.

The initialisation can be done inside a plug-in by calling the graph.
The above function takes over the expertise to do so correctly.
The input data is copied to output to avoid any conflicts during repeated
calling the graph. Even though by explicitely setting the input data as the
processing data for each filter allowes to work on one single data field.

The other option where to initialise completely from outside. This would
easily break or conflict with the high rate of expertise a plug-in is allowed
to have.

Image reading does still not work as expected.
(cherry picked from commit 1047653c0e193dc9b8f095c1fd052c8c4fbbc65e)