* [alpha]: oyFilterRegistrationToText supports masks
.. which allows to select multiple element levels from the registrations string.
oyPixelAccess_Create() API changed in order to attach additional data
from the current filter node to a ticket. The data option must be in the
"////resolve" registration range and will be transported through the filter
calls. A solution with a placeholder similiar to the hints in profiles or
even a text format seems a better choice. The transportation through the graph
has too much chances to fail.
(cherry picked from commit e07f3f13055152f63cef3def0585dcadec8e3731)