* [alpha]: define oyOptions_SetDriverContext()
The function shall take a driver context and merge this into options for
passing them to a device backend. The pointer can be of
* text (Oyranos options xml or text or some other driver_context_type)
* a binary blob - for easy and cheap serialisation

For example it is possible to pass a PPD or Gutenprint printrc.
The instructions how pass well behaving informations, which can the best
integrated into a ICC device + driver settings tag has not yet been specified.

This could include to parse and process a known driver_context inside the
device backend to maintain that work closer at the driver level.