* [alpha]: add oyProfile_DeviceGet() and ...
oyProfiles_DeviceRank().
oyProfile_DeviceGet() reads 'pddt', 'dmnd' and 'dmdd' into a
oyConfig_s::backend_core object.
oyProfiles_DeviceRank() performs a sort on a existing profile list
according to a given device. This allowes to put high device/driver ranks on
the top and sorted out ones on the end of a profile list.