* [and]: add non allocating string functions
This reduces the requests for memory allocation. Strings are seen by their
start pointer and length.
The usage of the length in the first one or four bytes would make this
even more compact. The price is a additional struct to hold the representation.
The current solution allowes to access strings in the usual C style with the
additional convention to use the string length as second argument. As this
construct is limited to some places for optimisations, it seems appropriate.