* [and]: switch memory pool off
The main problem is the constant work to be done in oyDeAllocateFunc_().
It does not matter wether quick small chunks or many cycle demanding chunks
are allocated. They are booth identified and must be checked by a single
pointer type. Therefor the near OS level memory allocation is too low.

It makes much more sense to store similiar sized blocks or objects. This is
already done with the oyObject_s type as the most frequently used type.

More optimisation might go into avoiding of computations like for hash sums.
If the registration matching is a real bootleneck there is still potential to
reduce the needed loops.