* [oyra]: add "filename" option
file_read and file_write attributes are a minimal means to see if the
filter is a file I/O filter. Better would be some way to tell that a it can
handle certain file types and delivers textual and preview information.
oyCMMDataTypes_s would almost fit but has some special meaning to filter
creation as exposed in oyCMMDataScan_f.
A new specialised type or generalisation of oyCMMDataTypes_s?