* [test]: support the "supported_devices_info"
a option emitted by device backends with only virtual devices during the "list"
call. The output is sorted to display manufacturers and models formated.

Each string in the list represents one manufacturer, which is present in the
first line of that string. The following lines contain each one model
separated by linebreak.