* [alpha]: add oyConfigs_Modify_f to oyCMMapi8_s
The new entry point shall provide existing oyConfig_s informations to modules.
Previously is was only supported to pass a "device_name" option to a module in
e.g. oyDeviceAskProfile(). This caused some essential loss of information.

The requirement of the SANE module to obtain device context information
in order to work properly with a user side context is the most demanding
reason for this change.