* [CUPS]: add local installation of a server profile
When a custom profile is found from the server, the CUPSgetProfiles is now
able to properly download the custom profile into one of Oyranos' ICC search
paths ('$HOME/.config/color').  While it seems that Oyranos itself is able
to successfully take in a blob of this custom profile,  the custom profile
doesn't seem to show up when I do a oyDeviceGetProfile() in KM (it returns
an sRGB profile instead).