* [alpha]: remove oyCMMapi_s::oyCMMCanHandle_f
The oyCMMCanHandle_f is almost obsolete and will be removed from most
module APIs except for oyCMMapi3_s - the profile tags modules. These
where the only using this capabilities. All other filters has moved this
functionality to other functions.

oyCMMapiBase_s will be merged with oyCMMapi_s. This simplifies the API
inheritance, if that term is allowed in a C project, of the modules.

oyCMMapi_s obtains a new member: int32_t          last_module_api[3];
This shall contain the version of the last module API of Oyranos used
during development of that filter. It is simply the actual development
version, e.g. at the moment {0,0,10} for Oyranos v0.0.10. With this new
member it will be possible to spot API incompatibilities in modules at
runtime, which otherwise easily can create crashes.

oyCMMapi1_s, the 1 dimensional pixel API, and oyCMMapi2_s, the monitor
functions, have as well obsoleted by the new filter and device module
architecture. As they are no longer used it seems appropriate to remove
them at the current state.