* [xforms]: move labels
Planing:
The next is to decide, who is responsible for profiles listings - Oyranos core
or the modules oyCMMuiGet() member. The later seems easier with the resulting
XFORMS becoming more final. This follows the i18n handling.
The core could do as well but needs more interfaces to pass options and parse.
In the end many modules have direct access to the used object types and data.