* [core]: add oyGROUP's to atstic options
This helps greatly in building XFORMS output.
The options can be obtained with oyWidgetListGet() and oyGROUP items come right
before the belonging items.
This is a alternative method to find the dependencies. The new method is much
easier to serialise.

The old method was to look in the oyOption_t_::category member and then parse
the already build tree and place the widget to the right place.