* [example]: add options editing in image_display
Ctrl-E or right mouse click should access a icc filter options dialog.
The options are passed as xml data model in a intermediate file. The XFORMS
layout is rendered by the external and experimental "oyranos-xforms-fltk"
application.

Missed is so far a signaling of the changed options to the
oyConversion_Context_s or the oyFilterNode_s.
A complete observation of changing options is troublesome.
A parallel damage, as now tried, is too much user and core work.
Idealy the options object should tell itself its observers about changes.
... a new oyObject_s property?