* [alpha]: add signal API
The system implements a very generic concept on the object level.
The flexibility is dangerous high.

Signals can be very useful to send messages where other means do not exist,
e.g. to interessted observers of a multiple referenced oyOption_s object.
This way the state of depending objects can change.

Inside the graph the signal concept is not adequate. The current implementation
misses  a real traversal of  the graph  to bring  oyFilterPlug/Socket_Callback
messages to the source of a call.  The  current  way is  to give return values
other than zero. Each filter has to check for the return value of the previous
filter and update its own state and depending contexts. The return value is in
danger to get lost. A alternative traversal  concept would require to register
a per module callback for doing the appropriate changes in the filter. A  good
time to finish that implementation.