* [oforms]: new oyUiHandler_s behaviour
The oyUiHandler_s have to parse themselfe the libxml2 structures. The Oyranos
option objects can not be the same time node and have attributes. So a mapping
to Oyranos objects was not only redundant, at least that would have given a
libxml2 abstraction, but it was a bad match to the XFORMS structure.

Oyranos does now only parse XFORMS for finding the interessting nodes and
passes them to the oyUiHandler_s. Some helper functions are provided to write
the code in a more compact manner. The oyUiHandler_s do only get the
xmlNodePtr from libxml2.

oyParseXMLNode_, the previous libxml2 to Oyranos object convertor,
is now obsolete.