* [alpha]: add oyCallback_s object
The object defines a generic structure for to be agreed conventions. The
callback can be casted to a more appropriate type. Arguments depend on the
convention. It will be good to decribe the convention. Perhaps through a
text string or even a name.

oyCallback_s can be used to pass a callback and according data to objects,
which shall be executed on agreed conditions. There is not much of type checks.