* [alpha]: change oyCMMui_s and oyCMMapi8_s
oyCMMui_s obtains a dynamic function for providing translated texts. This is
used to translate especially the category entry, which must be typically a
static string to keep the hierarchy clean. However for UI purposes a translated
version is prefered for display.

oyCMMapi8_s can with the new oyCMMui_s::ui and oyIcon_s::icon members provide
more informations to better satisfy KDE's KolorManager.

oyCMMInit_f is changed to accept a filter argument. This might be used in
further versions to extend the filter capabilities by dynamic registration.
In opposite, currently all properties must be expressed through the data
structures, which leads to frequent API breaks for unstable parts of the API.
The changed oyCMMInit_f function type allowes for a ABI compatible development.