* [oicc]: support proofing options
... especially "proof_soft", "proof_hard", "profiles_simulation".
The "proof_soft" option is only used in case the filter
"//imaging/display" is present in the graph. In case one of the above optins
apply and is not zero, the oyPROFILE_PROOF default profile will be used. That
is a setting from the CM panels default profiles.

The presence of profiles in "profiles_simulation" overrides all other options
and no further action is performed. The CMM will then pick and apply the
profile from "profiles_simulation".