* [exmpl]: support 0.4 DRAFT2 in colour_desktop
The "ICC Profiles in X Specification 0.4 DRAFT 2" is visible on ColourWiki and
in parts on the OpenICC email list. It allowes to change the _ICC_PROFILE(_xxx)
on the fly to move to the newly suggested _ICC_DEVIC_PROFILE(_xxx).
The _ICC_PROFILE(_xxx) atom is seen as document colour space and set typical
to sRGB (default in oyASSUMED_WEB). Pre 0.4 applications still work even if not
supporting larger than sRGB colour spaces. E.g. Inkscapes CM is fine after
these changes.

What is missed are to handle a newly set _ICC_PROFILE(_xxx) profile during a
session as described above. To get the old behaviour all has to be changed back.
This should be checked for switching the compiz plugin off.

Due to some problems with Tomas code for private data the private data is now
handled by a Oyranos cache mechanism. All pointers get a hash text and are
stored and obtained from a Oyranos cache.
This later changes where especially made for compatibility with the default
shipped compiz-0.7.8 for my actual developer system openSUSE-11.2 . Later
versions have to be tested.