* [exmpl]: bugfixes for 0.4 DRAFT2 colour_desktop
The changes allow to set v0.3 profiles and the compiz plugin taking care for
them to be honoured and the v0.4 DRAFT2 tags being updated.

The ignore profile list is not really nice. However it works ok.

I am not shure if the hasScreenProfile() test could be added into the oyConfig_s
device as a simple option upon setup. But it would be quite specialised.
The colour_desktop plugin shall not need too much of the knowledge of Oyranos'
oyX1 module. So hasScreenProfile() is better visible to other implementors.
0.4 DRAFT2 support relies havily on the colour server implementation. Its good
to have most stuff here except the v0.3, EDID and device profile DB support.

EDID propertiy events are no longer monitord. Not shure if this would be good to
support here as it is clearly v0.3 stuff.