* [alpha]: signals for changed options
Previously all touched options where signaled. The desired behaviour is to
only send signals for new key/value pairs. Thus the content must be checked.

Issues
It is not as easy with changing objects. There could the hash be used.
Currently only pointers of blobs and oyCMMptr_s' are checked. So unchanged
profiles would currently send a signal.