* [test]: fix test2 "oyOption_s integers"
The additional checks for already available values in a integer list fixed one
bug in oyOptions_SetFromInt(). The test2 code is now corrected for this case.
Strange that I did not see this before. Probably too few documentation.

Its possible that a bug is still around as the test does not cover the cases
where the changes to oyOptions_SetFromInt() do not take effect, namely setting
a new value.