First try to make the generator more clever.
Now old files in "src/" do not get deleted, but only updated.
Current status:
o New source files will be created if not present.
o If template is more recent, source file will be updated.
o If there is a change in the source file itself we only print a warning
o If contents of new - old file match operation is skipped.

Todo items:
o When there are changes in the source file, propagate them in the
template and rerun the renderer.
o Save the filestamps of all files and abort when both source and
template have been updated.