Separate the actual code from template code
1. All hand-written C code should be in sources/ directory.
2. The rest code creating templates inside the templates/
directory.

That should help towards the following goals:
(a) To make a change in the object model, edit some files
in templates/
(b) To update the insides of hand-written code, edit the *.c
files in sources/
(c) In order to add a new class, add the appropriate files
in sources/ and run the code-generator. Any needed metadata
can be extracted by the generator from the Doxygen tags
inside the code, like @extends, etc.
(d) The generated code in API_generated/ will be:
    i) read-only template-generated and protected by appropriate comments.
    ii) editable, and clear fromm comments from which file inside
        sources/ it originates.