[templates] Rearrange all templates to fit previous commit
* Every grantlee variable belongs to the context of the currently
rendered template, although it might belong to another included
or inherited parent template.
* So, when the source files are included, we use template variables
to hold their names and we get the correct file name for the
current rendered class.
* The only exception is the names of the source files that hold
the inherited classs members, where they need to be in the context
of the parent template.

This simplifies the Class* templates a lot!

Additionally, the Class* templates are identical for both
"opaque pointer" and "hidden struct" implementations.
Just food for thought :-)