* [modules]: CMM_NICK is mostly implementation detail
So that lcms and lcm2 are not differnt standards. They both implement
icc.colour, the one which is extented by the oicc module.
And _lcms or respectively _lcm2 is now a implementation attribute.
The same scheme applies often to the other modules.

An exception is oyIM currently, as there is no other implementation to be
expected. At least there will be an other name required for that to be
registred.