* [alpha]: add oyFilterRegistrationModify()
The new API supports stripping out of attributes to registration strings.

Many APIs expect to call not into double installed modules, but to apply to
complementary modules with distinct feature sets. So listing all display
devices means that Xorg devices shall be mentioned and if a specialised driver
can find even more or alternative ones these shall be added to the list.
This is used now for many filter APIs and should improve the situation of
parallel installed modules.