* [lcm2]: support oyFLOAT
... by adding cmsSigDToB0Tag inside abstract gamut profile and
enabling oyFloat in several places.

Use cmsCreateProofingTransformTHR and parallelise LUT sampling with
gamutCheckSamplerFloat() and gamutCheckSampler16().

Data in gamutCheckSamplerFloat() is scaled to match the TYPE_Lab_FLT pixel
format from the lcms transform.

The text tag is fixed by using cmsMLU.

The float pipeline D2B0 consists of a
linear segmented curve, the lut, and a linear segmented curve.
Not shure if the curves can be omited.

The A2B0 pipeline is similiarily contructed but due to the allowed
arrangements in the mAB type.