* [API]: remove OY_DEFAULT_RENDERING_HIGH_PRECISSION
Its not used anywhere know, so the assumption is made to strip this together
with oyBEHAVIOUR_RENDERING_HIGH_PRECISSION and
oyWIDGET_RENDERING_HIGH_PRECISSION will cause NO harm.

Instead the option is moved to precalculation in 'lcms' and 'lcm2'. Its an
individual option not a general considered CMM one.

Visually affected are the image_display ICC module context dialog (CTRL-e)
and oyranos-config-fltk in the Rendering Options tab.