* [alpha]: add oyConversion_Correct() sees flags
The selection of advanced options where previously hardcoded in
oiccChangeNodeOption().
However it makes sense to incluence this selection through a
parameter. This is done in oyConversion_Correct(). The oyCMMapi9_s
iCMM API is changed to pass this parameter through to
oiccChangeNodeOption().

There are poossibly more alpha APIs in need of this parameter.
Currently the resolving of the options can happen at several
times. This should possibly happen earlier to influence the
options set in the oyFilter_s object in fewer points.