* [oform]: use oyranos-config-fltk i18n init
Not shure if it works with osX bundles.

Fl_Help_View is packed in a group, but does not seem to turn on the desired
line wrapping.

Appending "</p>" to the help view string in callback_help_view() helps
displaying too long text. It must be scrolled. Still line wrap is better.
May ask on the FLTK list.