* [templates]: add missed headers
The library has currently lots of unresoved dependencies.
I guess a core library for string functions, helpers and basic key wrapper
should be build. This can then be extended by the object module and further
by the configure object, ICC and colour functions and so on. So it becomes
really modular.
Not shure how to make pkg-config happy. Split in separate libraries? Windows
users will love it, huh.