* [core]: add oyStruct_GetInfo()
The user side function will help in getting additional messages out of
objects. The other ones are for object authors and more for internal usage.

An other possible API would be to just return the string from oyMessageFunc_
instead of already printing to the console. That would help where no further
formatting is needed, and just an other message output path is enough.