* [API]: no object reference in messaging APIs
This became necessary, while modularising in core, basic and various
object code files. The core API now no longer depends on a
typedef for oyStruct_s. The functionality is still the same.

One disadvantage might be the missed compile time checking for passing
only Oyranos objects as the context pointer.