* [alpha]: change the oyCMMGetText_f API
For runtime generated plug-ins the context must be passed to provide the
requested data. In C plugins this data is statical to the already defined
function. The function in a script plug-in will be only a generic one. This
script plugin function can not know for which object it shall return the data.
The new object argument solves this problem.

This is a API break and version 0.3.0 is the right one for that to do.