* [templates]: trace inheritance in oyObject_s
The parent_type_ is replaced by parent_types_. The later is a list
containing all inherited object types of the belonging oyStruct_s object.
This allowes to detect run time type checking to inherited objects
during object casts.

Thus inherited functions will be useable to a derived object. That is
useful in interfaces which might be extended by certain classes like in the
module APIs. E.g. oyConnector_s interface/functions can be applied to a
oyConnectImaging_s class, which is passes as a oyConnector_s type.