* [templates]: make contructor obligatory
... in Base_s_.c. Previously the contructor code was enclosed by a
{% block %} marker. It forced replacement in case of alternative code for
the according marker. This is the case for customConstructor,
copyConstructor and customDestructor block in the templates/BaseList_s_.c
file. But we want the BaseList derived objects containing the same constructor
code like the template plus some additional injections. The later is
reached by providing the three oyOptions_XXX__Members() functions.

The oyOptions_XXX__Members() functions are a requirement for new classes
anyway in the CLASSprivate_custom_definitions.c file.