* [exmpl]: add Oy_Fl_Box::observeICC()
This fixes the use of a wrong pointer. To better debug such issues the
osPointer_s is used instead of oyBlob_s to provide specific type informations
about non Oyranos objects like Oy_Fl_Box. The later is specific to the FLTK
Oyranos wrapper and should not populate the Oyranos object system.