* [test]: Region of Image conversion
The new tests show Image colour conversions. One checks accessing a region of
interesst. This later test resembles image_display's region access for moving
content around.

The conversion shall only read the RoI and write to the described target region
of the requested size. Currently its writes the whole image, which is bad
for repeatedly running the conversion in chunks.