* [alpha]: oyPixelAccess_s::start_xy is now double
As the coordiantes are now relative to the output images width we need
floating point precission.

oyConversion_RunPixels() and oyConversion_ChangeRectangle() args are changed to
double as well.

The oyPixelAccess_CalculateNextStartPixel() is removed. This iterator is
not much used in Oyranos and should be implemented by specialied iterators.
This can happen very conveniently in C++ on users side.