* [alpha]: strip namespace from "serial" meta key
This will help in further processing the serial device property.
manufacturer, model and serial are the base device properties used in a
generic fashion for all device classes.

The kind of implemented stripping is a quick and dirty way. The alternative
would be to collect all device dependent prefixes from the oyCMMAPI8_s modules
and strip them away from each key prior to comparision with the actual key.
That way the "serial" key would be more correctly found at the expense of more
programming and computational overhead.