* [oyRE]: accept DNG in is_raw()
oyRE has a function called is_raw() to check if a given image is a raw file.
Is this needed? When must a image be rejected?
I found DNGs of Exiv2::ImageType 4, which can be rendered by libraw.

What is best to to here? Do we need to exclude certain images types?

The according changes for the DNGs are in the end of the atached patch.

http://www.exiv2.org/doc/namespaceExiv2_1_1ImageType.html