* [alpha]: support module search for generic objects
oyStruct_GetTextFromModule implements this and is hooked into
oyStruct_GetTextFromModule_p. That completes this porting issue in the
oyStruct_s template.

oyCMMsGetFilterApis_() accepts now no registrations like for unknown objects.

oyProfile_FromStd() checks the profiles file name before returning a profile.
This fixes a bug, where the default profile was changed on runtime, but the
oyProfile_FromStd() returned always the first returned profile, which was wrong.