* [conf]: compile with debug macros by default
The --disable-debug configure option is added to disable the feature.
--enable-debug is ignored.

We want debug macros build in, in order to ask users to use the
OYRANOS_DEBUG=2 command line macro and enable detailed output about
the actual work and state inside the library. Otherwise users have to
recompile, which is not easy enough.