* [exmpl]: selectively set output bit depth
With the new API it is possible to tell the graph early what output can be
used for a loaded image.

So the GL based Shader box can now offer 16-bit at minimum and float. The
image is therefore immediately loaded to obtain the according bit depth
information from. Depending on the data type we handle different texture
buffer sizes and call the glDrawPixels command with the according argument.

At the moment image import and drawing are aligned. The drawing code could
even detect the bit depth to correctly call glDrawPixels.