* [sources]: fix compile warnings
The deallocate functions are used for certain object types.
Putting them inside the template is fine.

The CMM object types are not yet correctly implemented. So we generally
do no need to expose warnings here. We want to implement the accessor
functions later, which are needed for dynamic allocation of these module
object types.