* [sources]: add oyConfig_GetOptions()
The oyConfig_GetOptions() gives access to low level configuration data.
Currently the different Options are used to describe source and purpose.
That is basically meta data to meta data. However that meta data is
needed to be available for e.g. commit the right set to the persistent DB.
The source argument maps to the currently ifixed available options and
can later be extended.

add oyRankMapCopy() and oyRankMapRelease() helpers

rename RankPad -> RankMap