* [sources]: return references objects
Referenced objects are the default mode to handle object returning. So we
remove a few exceptions for non non referenced objects, which caused
crashes in test2.

The following change affects only internal code.
Rename some internal variables. The idea to have a variable to a
external object type and a variable pointing to the adress of the first
variableto is a nice idea. However, it needs lots of casts like: (*var_) .
Due to casting it gives in some situations to few certainty, which variable
is aktual used. Given that there where soem errors in this regards, I assume
it is better to have a single variable pointing normally to the internally
complete object type for direct member manipulation and casting on demand
to the external object type. The other wa is of course as well possible,
by declaring the variable to the external object type and cast on demand
to the internal object type. Just two variables are redundant and
somewhat complicated.

What would help with distinguishing of external and internal wariables is
a following underscore for complete types: var_ .

Some crashes are fixed by the variable renaming.