* [core]: more caching of module accesses
The search for meta API's, module lists and single modules is now
cached. Caching has of course side effects. But each application follows
a typical request pattern, which will be seen inside the cache. A server
for arbitrary combinations of request parameters will not work nicely.
If that is to be used either a intelligent cache release or some
adjustment API's/variables need to be implemented.

oyCMMCacheListGetEntry_ is relatively early verbose with oy_debug > 0 .