* [lraw]: generate ICC profile from camera matrix
That feature is similar to on the fly ICC profile from EDID
generation.

The rendering should colour wise be similar to that of dcraw. The difference
is that we output into native camera space and omit conversion. In order
to work with a native camera space, we need to describe the colours with
a ICC profile and attach that to the oyImage object.

The DNG spec writes about AnalogBalance * CameraCalibration2 * ColorMatrix2.
From that the CameraCalibration2 appears missed in my test image.

The profile whitepoint is set to D65 as a that appears pretty common. But
I might be wrong and D50 is more correct.

This commit should fix wired colour appearance in image_display and of the
lraw plugin for CinePaint.