* [exmpl]: polish oy-file options
Options are much robuster parsed due to the tools macros.

Writing of PNG additionally to PPM is added.

Obtaining a ICC is added.

Obtaining OpenICC device JSON is added.

The adding of meta infromation to a profile is missed. A regression.
But not sure if that is really needed, as the OpenICC JSON can be
added as meta information through using the tool:
 oy-profile -w -j device.json