* [tools]: add oy-profile-graph --illuminant option
... with the A and D65 args for the two CIE standard lights.

USE_CLIPPER can be used to test the clipper code. But the current
implementation has quite some limitations and artefacts. We need here
a offset of a single non closed path, which is not supported. The
illuminant A curve looks after the offset very scattered. Perhaps the
algorithm is not robust for that case.

The code was fixed for compiling as C++.

draw_illuminant() abstracts the spectral curve drawing. The algorithm used
is pretty simple. It draws a line segment from the current to the next curve
point and sets on the end a half circle. The line segments are drawn with
a position dependent gradient like already for the --kelvin option.

Precission is neglible. The drawings lambda are aprocimations for illustration
only.

Code is adjusted for the oyLinInterpolateRampF32() API change.