* [lcm2]: add debug hash + oyDL ID to CMM context
The lcms CMM context uses a intermediate device link profile. The standard
ICC ID of this profile is attached to the CMM context oyPointer_s. This ID is
shown during DL creation in _Convert() and during conversions in _Run() when
oy_debug is on.

Additionally the hash value of the contexts XML description is added.
That hash value is used inside the core's cache.