* [source]: renew hash after call to module
Modules are free to do with their node what they see fit. So we can not rely
on a given hash to stay valid after calling into api7 for context preparation.
This patch renews the hash based on a context description after calling
api7 ContextToMem.

This fixes a bug introduced in #041ebeb9, where is was not possible to
obtain a transform without precalculation_curves being set after opening
a raw image by the lraw module. The lcm2 options dialog, was without any effect
for that option due to that wrong hash value in the CMS cache.