* [sources]: complete oyCMMapi4_Create()
A shorter function without the oyCMMapi_Set() stuff would be more elegant
in writing, at the risc of user errors. I think the current form is less
error proune, as we can not much do about checking the completeness of
the structure other than during create time in this function.

Alternatively a oyCMMapi4_CheckCompleteness() function could be added -
but is appears not nice either.