* [core]: fix printf formating for hashes
This commit fixes a bug where a leading zero in a four byte hash segment
resulted in missing of the leading number. The new formating forces to print
at least 8 letters and fill empty letters with '0'. That helps to get
equal results as with the md5sum cli tool. Affected are a some occurences
of ICC hash sums as shown in the oymd5icc tool.