* [core]: remove cmm parameter from internal module APIs
The functions never really used the CMM nick parameter to select
any module, except for the profile tag handling modules. But that
brought an big foot print to the internal API design. By removing
this parameter, the core can become much smaller, which affects
the required APIs and and code paths. Caching can be simplified too.