* [test]: display number of sub tests
Tests shall cover an API area and usage scenario.

The cound of Sub Tests show more the number of API calls or checks.

It would be nice to reach at least a 1:1 ratio of API calls and sub tests.
Currently the tests are mostly added when something breaks and that shall
not repeat. Sometimes test describe an desired API behaviour and the
API function is developed along that. Some extensive tests check for
syntax conformance in a single function like the registration. So the
presented tests are a mix of intentions.