* [core]: add internal oyDB_s object
it is not exported and merely a wrapper for Elektras KDB handler.

Rationale:
The DB is called many times in some code paths. Creating KDB is expensive.
oyDB_s allowes to store KDB without knowing much about its usage. oyDB_s
is not intented for export or to be copied. So it is not a full object and
needs not to be.