* [exmpl]: support rank map JSON with oy-test-device -k
Enable oy-test-device to build a complete oyConfig_s device from
a -j deviceJSON and a -k rank_mapJSON. The -j -k options combination
is able to substitute the -c device-class -d device-number combination.
oy-test-device can now be used to pass in user defined devices and
rankmaps to assign, remove. Additionally list matching profiles localy
and from Taxi DB.

Discussion:
The -k option is not much in final form. So it is not documented. Before
documenting it needs proper testing of adding a ICC profile to a JSON
device, changing the profile and getting it in the device and list results.