* [oyIM]: add more checks and debug info for desc tag
Give warnings on non standard errors. Account to ambiguity of late
clearification in the spec : alignment/strlen.

Add debug output for each parsed section in ascii, unicode and script
with code, provided count and string length. (The later is not tested with
script and code not available in ascii.)

pseq still is parsed, which is a good sign