* [exmpl]: persistently store scale factor and reload
Support automatic scaling for width, height and fit inside.

The symbolic scale factor for fit and friends needs to be preserved.
That requires a not so nice extra call Oy_Fl_Image_Widget::resetScale().

However the filter keeps simple and does not know about the symbolic stuff.
If it would have to handle, then it would have to change request sizes
in the job ticket. That can interfere with the filters caller intentions,
which is a risc and thus so far avoided.