* [core]: export some debug API's for the object tree
The API's allow for snapshot views. They do not care about allocations,
as they are for debugging only.

It would be cool to have a full call trace for each reference. But that would
need specialised API from glib, libunwind or similar. However the the provided
tree view can show referenced objects and helps spotting areas of for
releasing objects.