* [core]: move some string functions to oyjl
The string API's are used to create more powerful JSON functions and thus
need to be shared between oyjl users. However most string functions are
already implemented in Oyranos and can be safely moved to oyjl. The
Oyranos function names are cleaned up. The oyjl string functions are not
exported as is to prevend symbol irritations by linkers. It would be
preferable to import those functions by a common library: liboyjl ?